@@ -1,6 +1,6 @@
 [project]
 name = "fjsplib"
-version = "0.0.2a"
+version = "0.0.2"
 description = "Python package to read and write FJSP instances."
 authors = [
     { name = "Leon Lan", email = "l.lan@vu.nl"},
@@ -21,10 +21,8 @@ classifiers = [
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
 ]
-requires-python = ">=3.9"
-dependencies = [
-    "numpy>=1.26.0",
-]
+requires-python = ">=3.10"
+dependencies = []
 
 
 [project.urls]
@@ -33,13 +31,20 @@ Repository = "https://github.com/PyJobShop/FJSPLIB"
 
 [dependency-groups]
 dev = [
+    "numpy>=1.26.0",
     "pre-commit>=3.8.0",
     "pytest>=8.3.2",
 ]
 
 
-[tool.uv]
-default-groups = ["dev"]
+[build-system]
+requires = ["uv_build>=0.8.10,<0.9.0"]
+build-backend = "uv_build"
+
+
+[tool.uv.build-backend]
+module-name = "fjsplib"
+module-root = ""
 
 
 [tool.mypy]
@@ -62,7 +67,6 @@ select = [
     "F",  # pyflakes
     "I",  # isort
     "C",  # flake8-comprehensions
-    "B",  # flake8-bugbear
 ]
 ignore = [
     "E741",  # ambiguous variable name, needed for indexing
@@ -72,8 +76,3 @@ ignore = [
 
 [tool.ruff.lint.per-file-ignores]
 "__init__.py" = ["F401"] # ignore unused module imports
-
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
